move lazy errors to different file to prevent circle import
